feat(scroll result window): solution found
The incomplete scroll appears to be the result of url-insert happening
asynchronously - scrolling to point-max in the main code occurs too
fast.
Solution: add simple advice (sparql-max-scroll) to url-insert to
scroll window to max after url-insert finished.
Scroll should not affect any other use of url-insert, since it happens
only if the buffer-local variable sparql-overwrite-results is both
bound and nil.